refactor: split into a cargo workspace (core / standalone / libretro)
Restructure the single feature-flagged crate into a three-crate workspace so
the libretro core is produced with the file name RetroArch expects, directly
from Cargo, without any post-build renaming script.

Crates:
- native32emu-core   library, the platform-independent emulator engine
- native32emu        binary, the standalone minifb front-end (-> native32-emu)
- native32emu-libretro  cdylib named native32emu_libretro
                        (-> native32emu_libretro.{dll,so,dylib})

Why:
- The cdylib file name is the lib target name, so a dedicated libretro crate
  named `native32emu_libretro` makes Cargo emit the correctly named core with
  no rename step. The build-libretro.{ps1,sh} scripts are therefore removed.
- Dependencies are physically isolated: the libretro crate pulls in only the
  core engine and `log` (no minifb/rodio/clap), and the core engine only pulls
  in minifb/rodio when the standalone front-end enables its `standalone`
  feature.

Also in this change:
- Flatten the core module layer: src/core/* moved to the core crate root, so
  paths are `native32emu_core::emulator` instead of `native32emu::core::...`.
- Move native32emu_libretro.info under crates/native32emu-libretro/ next to the
  crate that produces the core.
- Update CI (ci/nightly/release/sonar) to build per-crate with -p / --workspace
  and to stage the libretro core without renaming; nightly and release now also
  publish the libretro core artifacts.
- Re-home cargo-husky into the core crate's dev-dependencies.
- Update README build/usage/structure sections.

No behavioral change to the emulator. Standalone and libretro both build, the
core's 175 unit tests pass, and clippy/fmt are clean.
