+++ .github/workflows/ci.yml
@@ -30,10 +30,10 @@ jobs:
         run: cargo fmt --all -- --check
 
       - name: Run clippy
-        run: cargo clippy --all-targets --all-features -- -D warnings
+        run: cargo clippy --workspace --all-targets --all-features -- -D warnings
 
       - name: Build
-        run: cargo build --release
+        run: cargo build --workspace --release
 
       - name: Run tests
-        run: cargo test
+        run: cargo test --workspace
