+++ Cargo.toml
@@ -1,40 +1,28 @@
-[package]
-name = "native32-emu"
+[workspace]
+resolver = "2"
+members = [
+    "crates/native32emu-core",
+    "crates/native32emu",
+    "crates/native32emu-libretro",
+]
+
+[workspace.package]
 version = "0.1.0"
 edition = "2021"
 description = "A Native32 game emulator written in Rust"
+license = "BSD-3-Clause"
 
-[lib]
-name = "native32emu"
-crate-type = ["lib"]
-path = "src/lib.rs"
-
-[[bin]]
-name = "native32-emu"
-path = "src/main.rs"
-required-features = ["standalone"]
-
-[dependencies]
+[workspace.dependencies]
+native32emu-core = { path = "crates/native32emu-core" }
 anyhow = "1"
 thiserror = "2"
 log = "0.4"
 rand = "0.10"
 image = "0.25"
-
-# Standalone mode dependencies (optional)
-clap = { version = "4", features = ["derive"], optional = true }
-minifb = { version = "0.28", optional = true }
-rodio = { version = "0.22", features = ["mp3"], optional = true }
-env_logger = { version = "0.11", optional = true }
-
-[dev-dependencies]
-cargo-husky = { version = "1", default-features = false, features = ["precommit-hook", "run-cargo-fmt", "run-cargo-clippy"] }
-tempfile = "3"
-
-[features]
-default = ["standalone"]
-standalone = ["dep:clap", "dep:minifb", "dep:rodio", "dep:env_logger"]
-libretro = []
+clap = { version = "4", features = ["derive"] }
+minifb = "0.28"
+rodio = { version = "0.22", features = ["mp3"] }
+env_logger = "0.11"
 
 [profile.release]
 opt-level = 3
