+++ crates/native32emu-core/src/actions.rs
