+++ crates/native32emu-core/src/content_loader.rs
