+++ crates/native32emu-core/src/des_constants.rs
