+++ crates/native32emu-core/src/error.rs
