+++ crates/native32emu-core/src/frame_player.rs
