+++ crates/native32emu-core/src/image_decoder.rs
