+++ crates/native32emu-core/src/input_handler.rs
