+++ crates/native32emu-core/src/save_manager.rs
