+++ crates/native32emu-libretro/src/libretro/callbacks.rs
