+++ crates/native32emu-libretro/src/libretro/constants.rs
