+++ crates/native32emu-libretro/src/libretro/logger.rs
