+++ crates/native32emu-libretro/src/libretro/types.rs
