+++ crates/native32emu/src/standalone/cli.rs
