+++ sonar-project.properties
@@ -2,7 +2,7 @@ sonar.projectKey=jiangxincode_Native32Emu
 sonar.projectName=Native32Emu
 sonar.organization=jiangxincode-github
 
-sonar.sources=src
+sonar.sources=crates
 sonar.exclusions=**/target/**
 sonar.sourceEncoding=UTF-8
